MDEV-23915 ER_KILL_DENIED_ERROR not passed a thread id (part 2)
Per Marko's comment in JIRA, sql_kill is passing the thread id
as long long. We change the format of the error messages to match,
and cast the thread id to long long in sql_kill_user.

+++ sql/sql_parse.cc
@@ -9123,7 +9123,7 @@ void sql_kill_user(THD *thd, LEX_USER *user, killed_state state)
     my_ok(thd, rows);
     break;
   case ER_KILL_DENIED_ERROR:
-    my_error(error, MYF(0), (unsigned long) thd->thread_id);
+    my_error(error, MYF(0), (long long) thd->thread_id);
     break;
   case ER_OUT_OF_RESOURCES:
   default:
